// frontend/src/client/types.gen.ts
export type UserRegisterWithCode = {
    email: string;
    password: string;
    full_name?: (string | null);
    verification_code: string;
};